pyproject.toml
[project]
name = "ca-api-codelab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.5.0",
    "google-cloud-geminidataanalytics>=0.3.0",
    "pandas>=2.3.2",
]
